# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
servers:
  - url: localhost:8080/api
    description: Sandbox server (uses test data)
paths:
####################################################
################# crud для задач ###################   
#################################################### 
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task' 
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'       
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{ID}:              
    patch:
      summary: Update task by ID
      tags:
        - tasks
      parameters:
        - name: ID
          in: path
          schema:
            type: integer 
            format: uint 
          required: true
          description: The ID of the task to update
      requestBody:
        description: The task to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task' 
    delete:
      summary: Delete task by ID
      tags:
        - tasks
      parameters:
        - name: ID
          in: path
          schema:
            type: integer 
            format: uint
          required: true
          description: The ID of the task to delete
      responses:
        '204':  
          description: Task deleted successfully

####################################################
############### crud пользователей #################
####################################################     
  /users:
    get:
      summary: Get all users
      tags:
        - users
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      tags:
        - users
      requestBody:
        description: The user to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'       
      responses:
        '201':
          description: The created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{ID}:
    patch:
      summary: Update user by ID
      tags:
        - users
      parameters:
        - name: ID
          in: path
          schema:
            type: integer 
            format: uint 
          required: true
          description: The ID of the user to update
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete user by ID
      tags:
        - users
      parameters:
        - name: ID
          in: path
          schema:
            type: integer 
            format: uint
          required: true
          description: The ID of the user to delete
      responses:
        '204':  
          description: User deleted successfully
    get:
      summary: Get all tasks some user by userID      
      tags:
       - users
      parameters:
        - name: ID
          in: path
          schema:
            type: integer 
            format: uint
          required: true
          description: The ID of the user to get all tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task' 


####################################################
############ дополнительные структуры ##############    
#################################################### 
components:
  schemas:
    # задача  
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint 
        content:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint  
    # пользователь
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint 
        email:
          type: string
        password_hash:
          type: string    